### 2026-03-06 #10150 @editorialbot
[PRE REVIEW]: OpenReservoirComputing: GPU-Accelerated Reservoir Computing in JAX
**Submitting author:** @Jan-Williams (<a href="http://orcid.org/0009-0005-4955-0411">Jan Williams</a>)
**Repository:** https://github.com/Jan-Williams/OpenReservoirComputing
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
