### 2026-04-24 f5ecb28
Trimmed paper and added DOIs
