### 2026-04-24 7cd64af
Fixed formatting for paper pdf generation
