+++ JOSS/paper.md
@@ -54,11 +54,17 @@ Because of ORC's functional approach in JAX, built-in and user-created models pr
 
 
 # State of the Field
-| | Language | GPU | Auto. Differentiable | Parallelizable/Vectorizable | Forecasting | Classification | Control | Continuous Time |
-|---|---|---|---|---|---|---|---|---|
-| **OpenReservoirComputing** | Python | $\checkmark$ | $\checkmark$ | $\checkmark$/$\checkmark$ | $\checkmark$ | $\checkmark$ | $\checkmark$ | $\checkmark$ |
-| **ReservoirPy** | Python | $\checkmark$* | $\times$ | $\times$/$\times$ | $\checkmark$ | $\checkmark$ | $\times$ | $\times$ |
-| **ReservoirComputing.jl** | Julia | $\checkmark$ | $\checkmark$ | $\checkmark$/$\times$ | $\checkmark$ | $\checkmark$ | $\times$ | $\times$ |
+| | **ORC** | **ReservoirPy** | **RC.jl** |
+|---|---|---|---|
+| Language | Python | Python | Julia |
+| GPU | $\checkmark$ | $\checkmark$* | $\checkmark$ |
+| Auto. Differentiable | $\checkmark$ | $\times$ | $\checkmark$ |
+| Parallelizable | $\checkmark$ | $\times$ | $\checkmark$ |
+| Vectorizable | $\checkmark$ | $\times$ | $\times$ |
+| Forecasting | $\checkmark$ | $\checkmark$ | $\checkmark$ |
+| Classification | $\checkmark$ | $\checkmark$ | $\checkmark$ |
+| Control | $\checkmark$ | $\times$ | $\times$ |
+| Continuous Time | $\checkmark$ | $\times$ | $\times$ |
 
 Table 1: Comparison of reservoir computing libraries across key features. ✓ indicates full support; ✗ indicates no support. \*ReservoirPy's GPU support is available via its JAX backend (v0.4.0+) but does not fully exploit JAX's functional programming model. *Parallelizable* denotes native support for parallel RC architectures as in [@pathak2018model] and *vectorizable* denotes native support for vectorization (e.g. `vmap`). {#tbl:comparison}
 
@@ -75,7 +81,7 @@ Other open source libraries for RC include Pytorch-ESN [@nardo2018pytorchesn] an
 
 # Software Design
 
-![ORC three-layer pipeline architecture. Each reservoir computer (RC) is decomposed into (i) an ``embedding'' function that lifts a low-dimensional signal $u_t$ to the reservoir dimension, (ii) a ``driver'' function that propagates the reservoir state, and (iii) a ``readout'' that maps back to a target $y$. For control and forecasting RCs, the target $y$ is typically $u_{t+1}$ (either in the presence of a forcing term or not) and for classification the target $y$ is a label. \label{fig:architecture}](../imgs/architecture_paper.pdf)
+![ORC three-layer pipeline architecture. Each reservoir computer (RC) is decomposed into (i) an ``embedding'' function that lifts a low-dimensional signal $u_t$ to the reservoir dimension, (ii) a ``driver'' function that propagates the reservoir state, and (iii) a ``readout'' that maps back to a target $y$. For control and forecasting RCs, the target $y$ is typically $u_{t+1}$ (either in the presence of a forcing term or not) and for classification the target $y$ is a label. \label{fig:architecture}](../imgs/architecture_paper.svg)
 
 
[…]
