### 2026-04-24 02b1f91
Fixed caption formatting in joss paper
