### 2026-04-29 #10150 @nkrusch
👋 Greetings @rkdan, @mbsuraj, @CDJellen, @ymahlau and @Saran-nns,

would you be interested in reviewing this submission for JOSS?
