### 2026-05-02 #10486 @nkrusch
> FYI, I will be on vacation the next week, it is going to take a while before I can start with the review

Thanks for letting me know, this is fine. The general expectation for JOSS is that the first pass of reviews to be completed within 4 weeks. First pass means reviewer has inspected all checklist items and reported outstanding issues (the review may not be complete after the first pass). Since the review phase started now, the first pass should by done by end of the May.
