### 2026-05-17 3db4cd6
fix zsh pip install compatibility by quoting extra dependencies
