### 2026-05-18 cb060e1
Added jax vmap/jit/autodiff tests for embeddings, readouts, and models, extending the tests from cjellen on drivers. In doing so, revealed CESN incompatibility with AD, added back compatible adjoint and max steps args to CRCForecasterBase and CESNForecaster so continuous CESN forecast is back differentiable
