### 2026-05-18 b512855
Fixed GRU param order and added associated tests
