### 2026-05-18 79360c2
Added train example, fixed typos
