### 2026-05-18 8042d6f
Larger pdf fig
