### 2026-05-18 85a81dc
Fixed typos in paper
