### 2026-05-18 d9ff293
Shortened paper
