+++ src/orc/__init__.py
@@ -1,4 +1,6 @@
-"""Components of models. Setting x64 on import."""
+"""Components of models."""
+
+import warnings
 
 import jax
 
@@ -13,7 +15,13 @@ from orc import (
     utils,
 )
 
-jax.config.update("jax_enable_x64", True)
+if not jax.config.jax_enable_x64:
+    warnings.warn(
+        "orc requires float64 precision. Enable it before importing orc with: "
+        "jax.config.update('jax_enable_x64', True)",
+        UserWarning,
+        stacklevel=2,
+    )
 
 __all__ = [
     "forecaster",
