### 2026-05-19 99678b1
Removed silent enabling of float64; added enabling of float64 to tests and added warning if float64 isn't enabled
