+++ src/orc/embeddings.py
@@ -7,7 +7,6 @@ import jax
 import jax.numpy as jnp
 from jaxtyping import Array, Float
 
-jax.config.update("jax_enable_x64", True)
 
 
 class EmbedBase(eqx.Module, ABC):
