+++ tests/forecaster/test_forecast_models.py
@@ -5,6 +5,8 @@ import jax.numpy as jnp
 import numpy as np
 import pytest
 
+jax.config.update("jax_enable_x64", True)
+
 import orc
 import orc.data
 
