+++ tests/data/test_integrators.py
@@ -2,6 +2,8 @@ import diffrax
 import jax
 import jax.numpy as jnp
 
+jax.config.update("jax_enable_x64", True)
+
 import orc
 import orc.data
 
