+++ tests/test_readouts.py
@@ -3,6 +3,8 @@ import jax
 import jax.numpy as jnp
 import pytest
 
+jax.config.update("jax_enable_x64", True)
+
 import orc
 
 
