### 2026-05-19 3142483
Formatting pass; updated pyprojec.toml to disable E402 on tests
