### 2026-05-19 6265cd9
Formatting pass; updated pyprojec.toml to disable E402 on tests
