### 2026-05-19 7a6a162
Changed float64 warning message and added float64 enabling to notebooks
