### 2026-05-19 a45aa4f
Updated README.md with PyPI installation instructions and enable float64
