### 2026-05-19 7b341e6
Stripped plt.show calls from built-in visualizations, updated README to reflect
