### 2026-05-19 d524fd7
Stripped plt.show calls from built-in visualizations, updated README to reflect
