### 2026-05-19 4524416
Incorporated @ymahlau paper suggestions
