+++ JOSS/paper.bib
@@ -139,3 +139,11 @@
   publisher={Springer},
   doi={10.1007/978-3-030-61616-8_40}
 }
+
+@inproceedings{wyder2025common,
+  title={Common Task Framework For a Critical Evaluation of Scientific Machine Learning Algorithms},
+  author={Wyder, Philippe Martin and Goldfeder, Judah A and Yermakov, Alexey and Zhao, Yue and Riva, Stefano and Williams, Jan P. and Zoro, David and Rude, Amy Sara and Tomasetto, Matteo and Germany, Joe and Bakarji, Joseph and Maierhofer, Georg and Cranmer, Miles and Kutz, J. Nathan},
+  booktitle={Neural Information Processing Systems 2025 Datasets and Benchmarks Track},
+  year={2025},
+  url={https://openreview.net/forum?id=pkFKjmUE3L}
+}
