### 2026-05-19 82c5440
Tweaking diagram
