### 2026-05-20 775b6e9
Added type checking with ty and corrected incorrect type hints; type checking will now be tested on pushes
