### 2026-05-20 98d1321
Docs color fix and image cleanup
