+++ mkdocs.yml
@@ -48,9 +48,10 @@ plugins:
             show_signature_annotations: true
             signature_crossrefs: true
   - mkdocs-jupyter:
-      execute: false
+      execute: true
       allow_errors: false
       include_source: true
+      timeout: 600
 
 markdown_extensions:
   - pymdownx.highlight:
