### 2026-05-20 f6c7eb9
Updated installation instruction in docs, not just readme
