### 2026-05-20 e8410b2
Fixed typo in installation instructions in docs. Added language to reference ty usage
