### 2026-05-20 94b5e13
Fixed typo in installation instructions in docs. Added language to reference ty usage
