### 2026-05-20 ef34894
Updated jit compatibility example
