+++ src/orc/forecaster/base.py
@@ -58,12 +58,12 @@ class RCForecasterBase(eqx.Module, ABC):
     driver: DriverBase
     readout: ReadoutBase
     embedding: EmbedBase
-    in_dim: int
-    out_dim: int
-    res_dim: int
-    chunks: int = 0
-    dtype: type = jnp.float64
-    seed: int = 0
+    in_dim: int = eqx.field(static=True)
+    out_dim: int = eqx.field(static=True)
+    res_dim: int = eqx.field(static=True)
+    chunks: int = eqx.field(default=0, static=True)
+    dtype: type = eqx.field(default=jnp.float64, static=True)
+    seed: int = eqx.field(default=0, static=True)
 
     def __init__(
         self,
