+++ src/orc/control/models.py
@@ -1,5 +1,6 @@
 """Discrete ESN controller implementation."""
 
+import equinox as eqx
 import jax
 import jax.numpy as jnp
 
@@ -46,9 +47,9 @@ class ESNController(RCControllerBase):
         Replace embedding layer.
     """
 
-    res_dim: int
-    data_dim: int
-    control_dim: int
+    res_dim: int = eqx.field(static=True)
+    data_dim: int = eqx.field(static=True)
+    control_dim: int = eqx.field(static=True)
 
     def __init__(
         self,
