### 2026-05-20 efee0fb
Updated jit compatibility example
