### 2026-05-21 f2571f5
Trigger paper regeneration
