### 2026-05-22 98d12d8
Implemented paper changes
