### 2026-05-26 5550f6a
Multi-sequence training of RCForecasters + formatting
