+++ src/orc/forecaster/train.py
@@ -25,6 +25,7 @@ def train_RCForecaster(
     initial_res_state: Array | None = None,
     beta: float = 8e-8,
     batch_size: int | None = None,
+    multi_sequence: bool = False,
     **force_kwargs,
 ) -> tuple[VariableRCForecaster, Array]:
     """Unified training function for reservoir computer forecasters.
@@ -41,14 +42,21 @@ def train_RCForecaster(
     model : RCForecasterBase
         Reservoir computer forecaster model to train.
     train_seq : Array
-        Training input sequence for reservoir, (shape=(seq_len, data_dim)).
+        Training input sequence for reservoir. Shape ``(seq_len, data_dim)``
+        for a single trajectory, or ``(n_traj, seq_len, data_dim)`` when
+        ``multi_sequence=True``. All trajectories must have the same length.
     target_seq : Array, optional
-        Target sequence for training reservoir, (shape=(seq_len, data_dim)).
-        If None, defaults to train_seq[1:].
+        Target sequence for training reservoir. Shape ``(seq_len, data_dim)``
+        for a single trajectory, or ``(n_traj, seq_len, data_dim)`` when
+        ``multi_sequence=True``. If None, defaults to train_seq[1:] (or
+        train_seq[:, 1:] for multi-sequence).
     spinup : int
-        Initial transient of reservoir states to discard.
+        Initial transient of reservoir states to discard per trajectory.
     initial_res_state : Array, optional
-        Initial reservoir state. If None, uses model.driver.default_state().
+        Initial reservoir state. If None, uses model.driver.default_state()
+        (tiled across trajectories when ``multi_sequence=True``). When
+        ``multi_sequence=True`` and not None, must have a leading trajectory
+        axis, i.e. shape ``(n_traj, ...)``.
     beta : float
         Tikhonov regularization parameter.
     batch_size : int, optional
@@ -56,58 +64,86 @@ def train_RCForecaster(
         regression. If None (default), processes all reservoirs at once.
         Use smaller values to reduce memory usage for large numbers of
         parallel reservoirs. Only used when readout.chunks > 0.
+    multi_sequence : bool
+        If True, treat train_seq (and target_seq) as a batch of trajectories
+        with a leading trajectory axis. Reservoir states from all trajectories
+        are concatenated (after spinup) before solving the regression. This
+        flag is a static boolean suitable for use as ``static_argnums`` when
+        wrapping with ``jax.jit``.
     **force_kwargs
         Additional keyword arguments passed to model.force() (e.g. ts=t_train
-        for continuous models).
+        for continuous models). When ``multi_sequence=True``, array-valued
+        kwargs must have a leading trajectory axis (e.g. ts of shape
+        ``(n_traj, seq_len)``).
 
     Returns
     -------
     model : RCForecasterBase
         Trained model.
     tot_res_seq : Array
-        Full sequence of reservoir states from teacher forcing.
+        Full reservoir state sequence from teacher forcing. Shape
+        ``(seq_len, ...)`` for single-sequence or ``(n_traj, seq_len, ...)``
+        for multi-sequence.
     """
-    if spinup >= train_seq.shape[0]:
+    seq_len = train_seq.shape[1] if multi_sequence else train_seq.shape[0]
+    if spinup >= seq_len:
         raise ValueError(
             "spinup must be less than the length of the training sequence."
         )
 
     if initial_res_state is None:
-        initial_res_state = model.driver.default_state()
+        default = model.driver.default_state()
+        if multi_sequence:
[…]
