+++ tests/control/test_control_models.py
@@ -370,6 +370,156 @@ def test_train_rccontroller_quadratic(dummy_control_problem_params):
     assert jnp.all(jnp.isfinite(U_controlled))
 
 
+##################### MULTI-SEQUENCE TRAINING TESTS #####################
+
+
+@pytest.fixture
+def lorenz_controller_segments():
+    """Three Lorenz63 trajectories with control sequences for multi-sequence tests."""
+    seq_len = 200
+    u0s = [
+        np.array([0.05, 1.0, 1.05]),
+        np.array([1.0, 0.5, 2.0]),
+        np.array([-1.0, 2.0, 0.5]),
+    ]
+    state_segs = jnp.stack(
+        [orc.data.lorenz63(tN=5, dt=0.01, u0=u0)[0][:seq_len] for u0 in u0s]
+    )  # (3, 200, 3)
+    time = jnp.arange(seq_len, dtype=jnp.float64).reshape(-1, 1)
+    control_segs = jnp.stack(
+        [
+            0.1 * jnp.sin(2 * jnp.pi * (i + 1) * time / seq_len * jnp.array([1.0, 1.5]))
+            for i in range(3)
+        ]
+    )  # (3, 200, 2)
+    return state_segs, control_segs
+
+
+def test_multi_seq_controller_n1_matches_single_seq(lorenz_controller_segments):
+    """n_traj=1 multi_sequence controller training matches single-sequence."""
+    state_segs, control_segs = lorenz_controller_segments
+    U, C = state_segs[0], control_segs[0]  # (200, 3), (200, 2)
+
+    ctrl = orc.control.ESNController(data_dim=3, control_dim=2, res_dim=200, seed=0)
+
+    ctrl_single, _ = orc.control.train_RCController(ctrl, U, C, spinup=20)
+    ctrl_multi, _ = orc.control.train_RCController(
+        ctrl, U[None], C[None], spinup=20, multi_sequence=True
+    )
+
+    assert jnp.allclose(ctrl_single.readout.wout, ctrl_multi.readout.wout)
+
+
+def test_multi_seq_controller_n1_explicit_target(lorenz_controller_segments):
+    """n_traj=1 multi_sequence with explicit target_seq matches single-sequence."""
+    state_segs, control_segs = lorenz_controller_segments
+    U, C = state_segs[0], control_segs[0]
+    train, target, ctrl_in = U[:150], U[1:151], C[:150]
+
+    ctrl = orc.control.ESNController(data_dim=3, control_dim=2, res_dim=200, seed=0)
+
+    ctrl_single, _ = orc.control.train_RCController(
+        ctrl, train, ctrl_in, target_seq=target, spinup=20
+    )
+    ctrl_multi, _ = orc.control.train_RCController(
+        ctrl,
+        train[None],
+        ctrl_in[None],
+        target_seq=target[None],
+        spinup=20,
+        multi_sequence=True,
+    )
+
+    assert jnp.allclose(ctrl_single.readout.wout, ctrl_multi.readout.wout)
+
+
+def test_multi_seq_controller_equiv_manual_concat(lorenz_controller_segments):
+    """Multi-sequence controller training matches per-trajectory training."""
+    from orc.utils.regressions import _solve_all_ridge_reg
+
+    state_segs, control_segs = lorenz_controller_segments
+    spinup = 20
+    beta = 1e-6
+    ctrl = orc.control.ESNController(data_dim=3, control_dim=2, res_dim=200, seed=0)
+
+    ctrl_multi, _ = orc.control.train_RCController(
+        ctrl, state_segs, control_segs, spinup=spinup, beta=beta, multi_sequence=True
+    )
+
+    default_state = ctrl.driver.default_state()
[…]
