### 2026-07-21 17db146
Updated equivalence test to check on outputs, not C matrices
