### 2026-07-23 660705f
Bumped tolerance for multi control sequence testing
