### 2026-07-23 cdc228a
Bumped tolerance for multi control sequence testing
