### 2026-07-27 b82b2b3
Prevent gh-pages history from accumulating on every docs deploy
mkdocs gh-deploy was appending a new full-site commit on each deploy,
which had grown the gh-pages branch to ~334MB of stacked build
artifacts. --no-history makes each deploy a single squashed commit.
