### 2026-07-27 3f0a43b
Migrate dependency management to uv
Adopt uv for environment/dependency management and clean up packaging cruft
that had accumulated in pyproject.toml.

Packaging:
- Switch build backend from setuptools+setuptools_scm to hatchling+hatch-vcs.
  hatch-vcs wraps setuptools_scm, so git-tag-derived version strings are
  unchanged (verified: 0.3.4.dev2+gc0e77dd2b before and after).
- Drop dead build requirements (wheel, numpy, cython) -- there are no C or
  Cython sources in the tree.
- Remove setuptools_scm from runtime dependencies; it is a build-time tool and
  nothing imports it, but every insta
[…]
