### 2026-07-27 b428921
Force uv-managed Python in CI to fix 3.12 segfault
The run-tests 3.12 job segfaulted (exit 139) while importing orc under
coverage, whereas 3.10, 3.11 and 3.13 all passed with identical dependency
versions -- `uv lock --upgrade` resolves to exactly the same versions as the
committed lock, so this was not an upstream regression.

The traceback showed the interpreter was /usr/lib/python3.12: ubuntu-latest
ships 3.12 as its system Python, so uv's default python-preference of "managed"
(prefer managed, fall back to system) used the system interpreter for that one
leg while downloading python-build
[…]
