### 2026-07-27 d7202df
Validate docs and packaging on pull requests; fix install docs
Add two pre-merge checks that were missing. docs.yml only ran on pushes to
main, so a broken example notebook was caught by the deploy job after merge --
leaving main red and the published site stale. publish.yml only runs on release,
so a packaging regression would first surface during an actual PyPI upload.

- docs.yml: add a pull_request trigger with a build-docs job that runs
  `mkdocs build --strict` and nothing else. Because mkdocs-jupyter runs with
  execute: true and allow_errors: false, this executes every example notebook
[…]
