+++ .github/workflows/docs.yml
@@ -1,11 +1,13 @@
-name: Deploy Docs
+name: Docs
 
 on:
   push:
     branches: [main]
+  pull_request:
+    branches: [main]
 
-permissions:
-  contents: write
+# Read-only by default; only the deploy job opts in to contents: write.
+permissions: {}
 
 # Use a uv-downloaded interpreter rather than the runner's system Python, so the
 # docs build matches the test matrix and local development.
@@ -18,8 +20,15 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  deploy:
+  # Pull requests: build only, no deploy and no write access. mkdocs-jupyter runs
+  # with execute: true and allow_errors: false, so this executes every example
+  # notebook and fails on a broken one -- catching it before merge rather than
+  # after, when the deploy job would be the first thing to notice.
+  build-docs:
+    if: github.event_name == 'pull_request'
     runs-on: ubuntu-latest
+    permissions:
+      contents: read
     steps:
       - uses: actions/checkout@v4
         with:
@@ -32,8 +41,32 @@ jobs:
           enable-cache: true
 
       # --no-default-groups drops the dev tools; --group docs still installs the
-      # project itself and its runtime deps, which the notebook execution needs
-      # (mkdocs.yml sets mkdocs-jupyter execute: true).
+      # project itself and its runtime deps, which the notebook execution needs.
+      - name: Install dependencies
+        run: uv sync --locked --no-default-groups --group docs
+
+      - name: Build docs
+        run: uv run mkdocs build --strict
+
+  # main: gh-deploy builds and publishes in one step, so this does not repeat the
+  # build job above (which would execute every notebook a second time). The two
+  # jobs are gated on different events and never run together.
+  deploy:
+    if: github.event_name == 'push'
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0
+
+      - uses: astral-sh/setup-uv@v9.0.0
+        with:
+          version: "0.11.30"
+          python-version: "3.12"
+          enable-cache: true
+
       - name: Install dependencies
         run: uv sync --locked --no-default-groups --group docs
 
